pyproject.toml
[project]
name = "teawish"
version = "0.1.0"
description = ""
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "dishka>=1.5.3",
    "fastapi>=0.115.12",
    "jinja2>=3.1.6",
    "orjson>=3.10.16",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.2",
    "python-multipart (>=0.0.20,<0.0.21)",
    "starlette-admin (>=0.14.1,<0.15.0)",
    "mcstatus (>=12.0.1,<13.0.0)",
    "aiofiles (>=24.1.0,<25.0.0)",
]

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.mypy]
ignore_missing_imports = true
follow_imports = 'skip'
warn_no_return = false

[tool.ruff]
line-length = 120

[tool.ruff.format]
quote-style = "single"
